split all functions out to new file
